mass_printing/requirements.txt
comtypes==1.4.13
et_xmlfile==2.0.0
lxml==6.0.2
numpy==2.3.4
openpyxl==3.1.5
pandas==2.3.3
PyPDF2==3.0.1
python-dateutil==2.9.0.post0
python-docx==1.2.0
pytz==2025.2
six==1.17.0
typing_extensions==4.15.0
tzdata==2025.2
